Creacion de carpeta operaciones
