+++ operaciones/division.py
@@ -1,0 +1,3 @@
+def division(nro1: float, nro2: float) -> None:
+    resultado = nro1 / nro2
+    print(resultado)
