+++ operaciones/multiplicacion.py
@@ -1,0 +1,3 @@
+def multiplicacion(nro1: float, nro2: float) -> None:
+    resultado = nro1 * nro2
+    print(resultado)
