+++ operaciones/suma.py
@@ -1,0 +1,3 @@
+def suma(nro1: float, nro2: float) -> None:
+    resultado = nro1 + nro2
+    print(resultado)
